requirements.txt
alembic==1.13.1
autopep8==2.2.0
bcrypt==4.1.3
cachetools==5.3.3
fastapi==0.111.0
isort==5.13.2
passlib==1.7.4
pydantic==2.7.3
pydantic_core==2.18.4
pydantic-settings==2.3.1
PyJWT==2.8.0
psycopg2-binary==2.9.9
sqlmodel==0.0.19
validator-collection==1.5.0
uvicorn==0.30.1